Add link to imgproxy package published in AUR for Arch Linux users
Fun fact: actually AUR doesn't contain packages and software is being build every time from the sources on end user's machine. So it is long and loud process. But it is as simple as package manager – you just check imgproxy and it will be installed for you along with libvips.