add option to exclude presets from signature
Currently, the entire URL path after the signature itself is
expected to be used for computing the signature. This is very limiting
in cases where an application needs to fetch the same image in
multiple formats from the proxy without having to create a different
signed URL for each format. In order to ensure that this facility
is provided without compromising security wherein - for example - a
hacker can fetch the full sized image given a thumbnail URL, we allow
applications to optionally exclude the presets part of the path from the
signature computation only when the IMGPROXY_ONLY_PRESETS option is
true. Thus, this will ensure that the most a hacker can do is fetch
a pre-defined set of presets given a URL that is signed using this
option - for applications that are OK assuming the associated risks.
This new bool option is called IMGPROXY_EXCLUDE_PRESETS_FROM_SIGNATURE.